Create EspAtDrv.cpp.patch
Hi Juraj,
I found out that the new EspressIF firmware v1.7x has the CIPSNTPTIME command but does lack the extension by LoBo firmware. Currently I'm using the original firmware and wanted to use getTime() as this is very very convenient. So I added just a couple of functions converting the Time String but not changing the original version by using an #ifdef .
It would be great if you could have a look on it and decide to add this to your wonderful library.
The file has been created with:  diff -u EspAtDrv.cpp.orig EspAtDrv.cpp  >patch.diff
Best,
Robert Sailer